v0.3.1
New mechanics, problem solve, code optimization
