+++ src/main/java/net/boybacks/superhardcore/recipes/items/craftable/Crystal.java
@@ -1,6 +1,7 @@
-package net.boybacks.superhardcore.craftingrecipes.items;
+package net.boybacks.superhardcore.recipes.items.craftable;
 
-import net.boybacks.superhardcore.craftingrecipes.handlers.onCraftingInventory;
+import lombok.Getter;
+import net.boybacks.superhardcore.recipes.handlers.onCraftingInventory;
 import net.boybacks.superhardcore.managers.ItemBuilderManager;
 import org.bukkit.Bukkit;
 import org.bukkit.Material;
@@ -13,10 +14,12 @@ import org.bukkit.inventory.ItemStack;
 
 import static net.boybacks.superhardcore.managers.ChatManager.fix;
 
-public class InfusedCrystal implements Listener {
+public class Crystal implements Listener {
+
+  @Getter static ItemStack crystal = new ItemBuilderManager(Material.EMERALD).setTitle("&d&lCrystal").addLoreLine("&8Custom Item").toItemStack();
 
   public static void onInventory(Player player) {
-    Inventory inventory = Bukkit.createInventory(player, 45, fix("&3&lRecipe: &d&lInfused Crystal"));
+    Inventory inventory = Bukkit.createInventory(player, 45, fix("&3&lRecipe: &d&lCrystal"));
     for (int i = 0; i < inventory.getSize(); i++) {
       if(inventory.getItem(i) == null || inventory.getItem(i).getType().equals(Material.AIR)) {
         ItemStack glassPane = new ItemBuilderManager(Material.WHITE_STAINED_GLASS_PANE).setTitle("&a").toItemStack();
@@ -27,22 +30,20 @@ public class InfusedCrystal implements Listener {
     ItemStack redstone = new ItemStack(Material.REDSTONE);
     inventory.setItem(10, redstone);
     inventory.setItem(12, redstone);
+    inventory.setItem(19, redstone);
+    inventory.setItem(21, redstone);
     inventory.setItem(28, redstone);
+    inventory.setItem(29, redstone);
     inventory.setItem(30, redstone);
 
-    ItemStack gunpowder = new ItemStack(Material.GUNPOWDER);
-    inventory.setItem(11, gunpowder);
-    inventory.setItem(19, gunpowder);
-    inventory.setItem(21, gunpowder);
+    ItemStack emerald = new ItemStack(Material.EMERALD);
+    inventory.setItem(11, emerald);
 
-    ItemStack blazePowder = new ItemStack(Material.BLAZE_POWDER);
-    inventory.setItem(29, blazePowder);
+    ItemStack amethyst = new ItemStack(Material.AMETHYST_SHARD);
+    inventory.setItem(20, amethyst);
 
     ItemStack crystal = new ItemBuilderManager(Material.EMERALD).setTitle("&d&lCrystal").addLoreLine("&8Custom Item").toItemStack();
-    inventory.setItem(20, crystal);
-
-    ItemStack infusedCrystal = new ItemBuilderManager(Material.EMERALD).setTitle("&d&lInfused Crystal").addLoreLine("&8Custom Item").addGlow().toItemStack();
-    inventory.setItem(24, infusedCrystal);
+    inventory.setItem(24, crystal);
 
     ItemStack exit = new ItemBuilderManager(Material.RED_STAINED_GLASS_PANE).setTitle("&c&lBack").toItemStack();
     inventory.setItem(40, exit);
@@ -55,7 +56,7 @@ public class InfusedCrystal implements Listener {
     if (event.getClickedInventory() == null) {
       return;
     }
-    if (event.getView().getTitle().equalsIgnoreCase(fix("&3&lRecipe: &d&lInfused Crystal"))) {
+    if (event.getView().getTitle().equalsIgnoreCase(fix("&3&lRecipe: &d&lCrystal"))) {
       event.setCancelled(true);
       Player player = (Player) event.getWhoClicked();
       if (event.getCurrentItem() == null) {
@@ -66,4 +67,5 @@ public class InfusedCrystal implements Listener {
       }
     }
   }
+
 }
