+++ src/main/java/net/boybacks/superhardcore/recipes/items/craftable/EmptySoulVial.java
@@ -1,6 +1,7 @@
-package net.boybacks.superhardcore.craftingrecipes.items;
+package net.boybacks.superhardcore.recipes.items.craftable;
 
-import net.boybacks.superhardcore.craftingrecipes.handlers.onCraftingInventory;
+import lombok.Getter;
+import net.boybacks.superhardcore.recipes.handlers.onCraftingInventory;
 import net.boybacks.superhardcore.managers.ItemBuilderManager;
 import org.bukkit.Bukkit;
 import org.bukkit.Material;
@@ -13,35 +14,34 @@ import org.bukkit.inventory.ItemStack;
 
 import static net.boybacks.superhardcore.managers.ChatManager.fix;
 
-public class Saddle implements Listener {
+public class EmptySoulVial implements Listener {
+
+  @Getter static ItemStack emptySoulVial = new ItemBuilderManager(Material.GLASS_BOTTLE).setTitle("&b&lEmpty Soul Vial").addLoreLine("&8Custom Item").toItemStack();
 
   public static void onInventory(Player player) {
-    Inventory inventory = Bukkit.createInventory(player, 45, fix("&3&lRecipe: &8Saddle"));
+    Inventory inventory = Bukkit.createInventory(player, 45, fix("&3&lRecipe: &b&lEmpty Soul Vial"));
     for (int i = 0; i < inventory.getSize(); i++) {
       if(inventory.getItem(i) == null || inventory.getItem(i).getType().equals(Material.AIR)) {
         ItemStack glassPane = new ItemBuilderManager(Material.WHITE_STAINED_GLASS_PANE).setTitle("&a").toItemStack();
         inventory.setItem(i, glassPane);
       }
     }
-    ItemStack air = new ItemStack(Material.AIR);
-    inventory.setItem(20, air);
-    inventory.setItem(29, air);
 
-    ItemStack leather = new ItemStack(Material.LEATHER);
-    inventory.setItem(10, leather);
-    inventory.setItem(11, leather);
-    inventory.setItem(12, leather);
+    ItemStack glassBottle = new ItemStack(Material.GLASS_BOTTLE);
+    inventory.setItem(20, glassBottle);
 
-    ItemStack copper = new ItemStack(Material.COPPER_INGOT);
-    inventory.setItem(19, copper);
-    inventory.setItem(21, copper);
+    ItemStack diamondNugget = new ItemBuilderManager(Material.IRON_NUGGET).setTitle("&b&lDiamond Nugget").addLoreLine("&8Custom Item").toItemStack();
+    inventory.setItem(10, diamondNugget);
+    inventory.setItem(11, diamondNugget);
+    inventory.setItem(12, diamondNugget);
+    inventory.setItem(19, diamondNugget);
+    inventory.setItem(21, diamondNugget);
+    inventory.setItem(28, diamondNugget);
+    inventory.setItem(29, diamondNugget);
+    inventory.setItem(30, diamondNugget);
 
-    ItemStack tripwireHook = new ItemStack(Material.TRIPWIRE_HOOK);
-    inventory.setItem(28, tripwireHook);
-    inventory.setItem(30, tripwireHook);
-
-    ItemStack saddle = new ItemStack(Material.SADDLE);
-    inventory.setItem(24, saddle);
+    ItemStack emptySoulVial = new ItemBuilderManager(Material.GLASS_BOTTLE).setTitle("&b&lEmpty Soul Vial").addLoreLine("&8Custom Item").toItemStack();
+    inventory.setItem(24, emptySoulVial);
 
     ItemStack exit = new ItemBuilderManager(Material.RED_STAINED_GLASS_PANE).setTitle("&c&lBack").toItemStack();
     inventory.setItem(40, exit);
@@ -54,7 +54,7 @@ public class Saddle implements Listener {
     if (event.getClickedInventory() == null) {
       return;
     }
-    if (event.getView().getTitle().equalsIgnoreCase(fix("&3&lRecipe: &8Saddle"))) {
+    if (event.getView().getTitle().equalsIgnoreCase(fix("&3&lRecipe: &b&lEmpty Soul Vial"))) {
       event.setCancelled(true);
       Player player = (Player) event.getWhoClicked();
       if (event.getCurrentItem() == null) {
