+++ src/main/java/net/boybacks/superhardcore/recipes/items/craftable/Heart.java
@@ -1,6 +1,7 @@
-package net.boybacks.superhardcore.craftingrecipes.items;
+package net.boybacks.superhardcore.recipes.items.craftable;
 
-import net.boybacks.superhardcore.craftingrecipes.handlers.onCraftingInventory;
+import lombok.Getter;
+import net.boybacks.superhardcore.recipes.handlers.onCraftingInventory;
 import net.boybacks.superhardcore.managers.ItemBuilderManager;
 import org.bukkit.Bukkit;
 import org.bukkit.Material;
@@ -13,35 +14,34 @@ import org.bukkit.inventory.ItemStack;
 
 import static net.boybacks.superhardcore.managers.ChatManager.fix;
 
-public class TotemOfUndying implements Listener {
+public class Heart implements Listener {
+
+  @Getter static ItemStack heart = new ItemBuilderManager(Material.FERMENTED_SPIDER_EYE).setTitle("&c&lHeart").addLoreLine("&8Custom Item").toItemStack();
 
   public static void onInventory(Player player) {
-    Inventory inventory = Bukkit.createInventory(player, 45, fix("&3&lRecipe: &e&lTotem of Undying"));
+    Inventory inventory = Bukkit.createInventory(player, 45, fix("&3&lRecipe: &c&lHeart"));
     for (int i = 0; i < inventory.getSize(); i++) {
       if(inventory.getItem(i) == null || inventory.getItem(i).getType().equals(Material.AIR)) {
         ItemStack glassPane = new ItemBuilderManager(Material.WHITE_STAINED_GLASS_PANE).setTitle("&a").toItemStack();
         inventory.setItem(i, glassPane);
       }
     }
-    ItemStack emerald = new ItemStack(Material.EMERALD);
-    inventory.setItem(10, emerald);
-    inventory.setItem(12, emerald);
 
-    ItemStack goldIngot = new ItemStack(Material.GOLD_INGOT);
-    inventory.setItem(28, goldIngot);
-    inventory.setItem(30, goldIngot);
+    ItemStack heartChunk = new ItemBuilderManager(Material.SPIDER_EYE).setTitle("&c&lHeart Chunk").addLoreLine("&8Custom Item").toItemStack();
+    inventory.setItem(10, heartChunk);
+    inventory.setItem(11, heartChunk);
+    inventory.setItem(12, heartChunk);
+    inventory.setItem(19, heartChunk);
+    inventory.setItem(21, heartChunk);
+    inventory.setItem(28, heartChunk);
+    inventory.setItem(29, heartChunk);
+    inventory.setItem(30, heartChunk);
 
-    ItemStack rippedHeart = new ItemBuilderManager(Material.FERMENTED_SPIDER_EYE).setTitle("&c&lRipped Heart").addLoreLine("&8Custom Item").addGlow().toItemStack();
-    inventory.setItem(20, rippedHeart);
+    ItemStack soulVial = new ItemBuilderManager(Material.GLASS_BOTTLE).setTitle("&b&lSoul Vial").addLoreLine("&8Custom Item").toItemStack();
+    inventory.setItem(20, soulVial);
 
-    ItemStack infusedCrystal = new ItemBuilderManager(Material.EMERALD).setTitle("&d&lInfused Crystal").addLoreLine("&8Custom Item").addGlow().toItemStack();
-    inventory.setItem(11, infusedCrystal);
-    inventory.setItem(19, infusedCrystal);
-    inventory.setItem(21, infusedCrystal);
-    inventory.setItem(29, infusedCrystal);
-
-    ItemStack totemOfUndying = new ItemBuilderManager(Material.TOTEM_OF_UNDYING).toItemStack();
-    inventory.setItem(24, totemOfUndying);
+    ItemStack heart = new ItemBuilderManager(Material.FERMENTED_SPIDER_EYE).setTitle("&c&lHeart").addLoreLine("&8Custom Item").toItemStack();
+    inventory.setItem(24, heart);
 
     ItemStack exit = new ItemBuilderManager(Material.RED_STAINED_GLASS_PANE).setTitle("&c&lBack").toItemStack();
     inventory.setItem(40, exit);
@@ -54,7 +54,7 @@ public class TotemOfUndying implements Listener {
     if (event.getClickedInventory() == null) {
       return;
     }
-    if (event.getView().getTitle().equalsIgnoreCase(fix("&3&lRecipe: &e&lTotem of Undying"))) {
+    if (event.getView().getTitle().equalsIgnoreCase(fix("&3&lRecipe: &c&lHeart"))) {
       event.setCancelled(true);
       Player player = (Player) event.getWhoClicked();
       if (event.getCurrentItem() == null) {
