+++ src/main/java/net/boybacks/superhardcore/recipes/items/craftable/InfusedCrystal.java
@@ -1,6 +1,7 @@
-package net.boybacks.superhardcore.craftingrecipes.items;
+package net.boybacks.superhardcore.recipes.items.craftable;
 
-import net.boybacks.superhardcore.craftingrecipes.handlers.onCraftingInventory;
+import lombok.Getter;
+import net.boybacks.superhardcore.recipes.handlers.onCraftingInventory;
 import net.boybacks.superhardcore.managers.ItemBuilderManager;
 import org.bukkit.Bukkit;
 import org.bukkit.Material;
@@ -13,10 +14,12 @@ import org.bukkit.inventory.ItemStack;
 
 import static net.boybacks.superhardcore.managers.ChatManager.fix;
 
-public class RippedHeart implements Listener {
+public class InfusedCrystal implements Listener {
+
+  @Getter static ItemStack infusedCrystal = new ItemBuilderManager(Material.EMERALD).setTitle("&d&lInfused Crystal").addLoreLine("&8Custom Item").addGlow().toItemStack();
 
   public static void onInventory(Player player) {
-    Inventory inventory = Bukkit.createInventory(player, 45, fix("&3&lRecipe: &c&lRipped Heart"));
+    Inventory inventory = Bukkit.createInventory(player, 45, fix("&3&lRecipe: &d&lInfused Crystal"));
     for (int i = 0; i < inventory.getSize(); i++) {
       if(inventory.getItem(i) == null || inventory.getItem(i).getType().equals(Material.AIR)) {
         ItemStack glassPane = new ItemBuilderManager(Material.WHITE_STAINED_GLASS_PANE).setTitle("&a").toItemStack();
@@ -24,23 +27,25 @@ public class RippedHeart implements Listener {
       }
     }
 
+    ItemStack redstone = new ItemStack(Material.REDSTONE);
+    inventory.setItem(10, redstone);
+    inventory.setItem(12, redstone);
+    inventory.setItem(28, redstone);
+    inventory.setItem(30, redstone);
+
+    ItemStack gunpowder = new ItemStack(Material.GUNPOWDER);
+    inventory.setItem(11, gunpowder);
+    inventory.setItem(19, gunpowder);
+    inventory.setItem(21, gunpowder);
+
+    ItemStack blazePowder = new ItemStack(Material.BLAZE_POWDER);
+    inventory.setItem(29, blazePowder);
+
+    ItemStack crystal = new ItemBuilderManager(Material.EMERALD).setTitle("&d&lCrystal").addLoreLine("&8Custom Item").toItemStack();
+    inventory.setItem(20, crystal);
+
     ItemStack infusedCrystal = new ItemBuilderManager(Material.EMERALD).setTitle("&d&lInfused Crystal").addLoreLine("&8Custom Item").addGlow().toItemStack();
-    inventory.setItem(10, infusedCrystal);
-    inventory.setItem(12, infusedCrystal);
-    inventory.setItem(28, infusedCrystal);
-    inventory.setItem(30, infusedCrystal);
-
-    ItemStack goldIngot = new ItemStack(Material.GOLD_INGOT);
-    inventory.setItem(11, goldIngot);
-    inventory.setItem(19, goldIngot);
-    inventory.setItem(21, goldIngot);
-    inventory.setItem(29, goldIngot);
-
-    ItemStack heart = new ItemBuilderManager(Material.FERMENTED_SPIDER_EYE).setTitle("&c&lHeart").addLoreLine("&8Custom Item").toItemStack();
-    inventory.setItem(20, heart);
-
-    ItemStack rippedHeart = new ItemBuilderManager(Material.FERMENTED_SPIDER_EYE).setTitle("&c&lRipped Heart").addLoreLine("&8Custom Item").addGlow().toItemStack();
-    inventory.setItem(24, rippedHeart);
+    inventory.setItem(24, infusedCrystal);
 
     ItemStack exit = new ItemBuilderManager(Material.RED_STAINED_GLASS_PANE).setTitle("&c&lBack").toItemStack();
     inventory.setItem(40, exit);
@@ -53,7 +58,7 @@ public class RippedHeart implements Listener {
     if (event.getClickedInventory() == null) {
       return;
     }
-    if (event.getView().getTitle().equalsIgnoreCase(fix("&3&lRecipe: &c&lRipped Heart"))) {
+    if (event.getView().getTitle().equalsIgnoreCase(fix("&3&lRecipe: &d&lInfused Crystal"))) {
       event.setCancelled(true);
       Player player = (Player) event.getWhoClicked();
       if (event.getCurrentItem() == null) {
