+++ src/main/java/net/boybacks/superhardcore/recipes/items/craftable/SacrificeDagger.java
@@ -1,6 +1,7 @@
-package net.boybacks.superhardcore.craftingrecipes.items;
+package net.boybacks.superhardcore.recipes.items.craftable;
 
-import net.boybacks.superhardcore.craftingrecipes.handlers.onCraftingInventory;
+import lombok.Getter;
+import net.boybacks.superhardcore.recipes.handlers.onCraftingInventory;
 import net.boybacks.superhardcore.managers.ItemBuilderManager;
 import org.bukkit.Bukkit;
 import org.bukkit.Material;
@@ -10,35 +11,42 @@ import org.bukkit.event.Listener;
 import org.bukkit.event.inventory.InventoryClickEvent;
 import org.bukkit.inventory.Inventory;
 import org.bukkit.inventory.ItemStack;
+import org.bukkit.potion.PotionType;
 
 import static net.boybacks.superhardcore.managers.ChatManager.fix;
 
-public class EmptySoulVial implements Listener {
+
+public class SacrificeDagger implements Listener {
+
+  @Getter static ItemStack sacrificeDagger = new ItemBuilderManager(Material.GOLDEN_SWORD).setTitle("&6&lSacrifice Dagger").addLoreLine("&8Custom Item").setDurability((short) 22).toItemStack();
 
   public static void onInventory(Player player) {
-    Inventory inventory = Bukkit.createInventory(player, 45, fix("&3&lRecipe: &b&lEmpty Soul Vial"));
+    Inventory inventory = Bukkit.createInventory(player, 45, fix("&3&lRecipe: &6&lSacrifice Dagger"));
     for (int i = 0; i < inventory.getSize(); i++) {
       if(inventory.getItem(i) == null || inventory.getItem(i).getType().equals(Material.AIR)) {
         ItemStack glassPane = new ItemBuilderManager(Material.WHITE_STAINED_GLASS_PANE).setTitle("&a").toItemStack();
         inventory.setItem(i, glassPane);
       }
     }
+    ItemStack air = new ItemStack(Material.AIR);
+    inventory.setItem(10, air);
+    inventory.setItem(19, air);
+    inventory.setItem(29, air);
+    inventory.setItem(30, air);
 
-    ItemStack glassBottle = new ItemStack(Material.GLASS_BOTTLE);
-    inventory.setItem(20, glassBottle);
+    ItemStack gold = new ItemStack(Material.GOLD_INGOT);
+    inventory.setItem(12, gold);
+    inventory.setItem(20, gold);
 
-    ItemStack diamondNugget = new ItemBuilderManager(Material.IRON_NUGGET).setTitle("&b&lDiamond Nugget").addLoreLine("&8Custom Item").toItemStack();
-    inventory.setItem(10, diamondNugget);
-    inventory.setItem(11, diamondNugget);
-    inventory.setItem(12, diamondNugget);
-    inventory.setItem(19, diamondNugget);
-    inventory.setItem(21, diamondNugget);
-    inventory.setItem(28, diamondNugget);
-    inventory.setItem(29, diamondNugget);
-    inventory.setItem(30, diamondNugget);
+    ItemStack stick = new ItemStack(Material.STICK);
+    inventory.setItem(28, stick);
 
-    ItemStack emptySoulVial = new ItemBuilderManager(Material.GLASS_BOTTLE).setTitle("&b&lEmpty Soul Vial").addLoreLine("&8Custom Item").toItemStack();
-    inventory.setItem(24, emptySoulVial);
+    ItemStack potion = new ItemBuilderManager(Material.POTION).setPotionType(PotionType.POISON, false, true).toItemStack();
+    inventory.setItem(11, potion);
+    inventory.setItem(21, potion);
+
+    ItemStack dagger = new ItemBuilderManager(Material.GOLDEN_SWORD).setTitle("&6&lSacrifice Dagger").addLoreLine("&8Custom Item").setDurability((short) 22).toItemStack();
+    inventory.setItem(24, dagger);
 
     ItemStack exit = new ItemBuilderManager(Material.RED_STAINED_GLASS_PANE).setTitle("&c&lBack").toItemStack();
     inventory.setItem(40, exit);
@@ -51,7 +59,7 @@ public class EmptySoulVial implements Listener {
     if (event.getClickedInventory() == null) {
       return;
     }
-    if (event.getView().getTitle().equalsIgnoreCase(fix("&3&lRecipe: &b&lEmpty Soul Vial"))) {
+    if (event.getView().getTitle().equalsIgnoreCase(fix("&3&lRecipe: &6&lSacrifice Dagger"))) {
       event.setCancelled(true);
       Player player = (Player) event.getWhoClicked();
       if (event.getCurrentItem() == null) {
@@ -62,4 +70,6 @@ public class EmptySoulVial implements Listener {
       }
     }
   }
+
+
 }
