+++ src/main/java/net/boybacks/superhardcore/recipes/items/craftable/SoulVial.java
@@ -1,6 +1,7 @@
-package net.boybacks.superhardcore.craftingrecipes.items;
+package net.boybacks.superhardcore.recipes.items.craftable;
 
-import net.boybacks.superhardcore.craftingrecipes.handlers.onCraftingInventory;
+import lombok.Getter;
+import net.boybacks.superhardcore.recipes.handlers.onCraftingInventory;
 import net.boybacks.superhardcore.managers.ItemBuilderManager;
 import org.bukkit.Bukkit;
 import org.bukkit.Material;
@@ -13,11 +14,12 @@ import org.bukkit.inventory.ItemStack;
 
 import static net.boybacks.superhardcore.managers.ChatManager.fix;
 
-public class DiamondNugget implements Listener {
+public class SoulVial implements Listener {
 
+  @Getter static ItemStack soulVial = new ItemBuilderManager(Material.GLASS_BOTTLE).setTitle("&b&lSoul Vial").addLoreLine("&8Custom Item").toItemStack();
 
   public static void onInventory(Player player) {
-    Inventory inventory = Bukkit.createInventory(player, 45, fix("&3&lFurnace: &b&lDiamond Nugget"));
+    Inventory inventory = Bukkit.createInventory(player, 45, fix("&3&lRecipe: &b&lSoul Vial"));
     for (int i = 0; i < inventory.getSize(); i++) {
       if(inventory.getItem(i) == null || inventory.getItem(i).getType().equals(Material.AIR)) {
         ItemStack glassPane = new ItemBuilderManager(Material.WHITE_STAINED_GLASS_PANE).setTitle("&a").toItemStack();
@@ -25,17 +27,23 @@ public class DiamondNugget implements Listener {
       }
     }
 
-    ItemStack coal = new ItemStack(Material.COAL);
-    inventory.setItem(29, coal);
-
-    ItemStack diamond = new ItemStack(Material.DIAMOND);
-    inventory.setItem(11, diamond);
-
-    ItemStack fire = new ItemBuilderManager(Material.ORANGE_STAINED_GLASS_PANE).setTitle("&a").toItemStack();
-    inventory.setItem(20, fire);
-
     ItemStack diamondNugget = new ItemBuilderManager(Material.IRON_NUGGET).setTitle("&b&lDiamond Nugget").addLoreLine("&8Custom Item").toItemStack();
-    inventory.setItem(24, diamondNugget);
+    inventory.setItem(11, diamondNugget);
+
+    ItemStack wildSoul = new ItemBuilderManager(Material.FEATHER, 1).setTitle("&b&lWild Soul").addLoreLine("&8Custom Item").toItemStack();
+    inventory.setItem(10, wildSoul);
+    inventory.setItem(12, wildSoul);
+    inventory.setItem(19, wildSoul);
+    inventory.setItem(21, wildSoul);
+    inventory.setItem(28, wildSoul);
+    inventory.setItem(29, wildSoul);
+    inventory.setItem(30, wildSoul);
+
+    ItemStack emptySoulVial = new ItemBuilderManager(Material.GLASS_BOTTLE).setTitle("&b&lEmpty Soul Vial").addLoreLine("&8Custom Item").toItemStack();
+    inventory.setItem(20, emptySoulVial);
+
+    ItemStack soulVial = new ItemBuilderManager(Material.GLASS_BOTTLE).setTitle("&b&lSoul Vial").addLoreLine("&8Custom Item").toItemStack();
+    inventory.setItem(24, soulVial);
 
     ItemStack exit = new ItemBuilderManager(Material.RED_STAINED_GLASS_PANE).setTitle("&c&lBack").toItemStack();
     inventory.setItem(40, exit);
@@ -48,7 +56,7 @@ public class DiamondNugget implements Listener {
     if (event.getClickedInventory() == null) {
       return;
     }
-    if (event.getView().getTitle().equalsIgnoreCase(fix("&3&lFurnace: &b&lDiamond Nugget"))) {
+    if (event.getView().getTitle().equalsIgnoreCase(fix("&3&lRecipe: &b&lSoul Vial"))) {
       event.setCancelled(true);
       Player player = (Player) event.getWhoClicked();
       if (event.getCurrentItem() == null) {
