+++ src/main/java/net/boybacks/superhardcore/recipes/items/craftable/vanilla/Saddle.java
@@ -1,6 +1,7 @@
-package net.boybacks.superhardcore.craftingrecipes.items;
+package net.boybacks.superhardcore.recipes.items.craftable.vanilla;
 
-import net.boybacks.superhardcore.craftingrecipes.handlers.onCraftingInventory;
+import lombok.Getter;
+import net.boybacks.superhardcore.recipes.handlers.onCraftingInventory;
 import net.boybacks.superhardcore.managers.ItemBuilderManager;
 import org.bukkit.Bukkit;
 import org.bukkit.Material;
@@ -13,31 +14,37 @@ import org.bukkit.inventory.ItemStack;
 
 import static net.boybacks.superhardcore.managers.ChatManager.fix;
 
-public class Heart implements Listener {
+public class Saddle implements Listener {
+
+  @Getter static ItemStack saddle = new ItemStack(Material.SADDLE);
+
   public static void onInventory(Player player) {
-    Inventory inventory = Bukkit.createInventory(player, 45, fix("&3&lRecipe: &c&lHeart"));
+    Inventory inventory = Bukkit.createInventory(player, 45, fix("&3&lRecipe: &8Saddle"));
     for (int i = 0; i < inventory.getSize(); i++) {
       if(inventory.getItem(i) == null || inventory.getItem(i).getType().equals(Material.AIR)) {
         ItemStack glassPane = new ItemBuilderManager(Material.WHITE_STAINED_GLASS_PANE).setTitle("&a").toItemStack();
         inventory.setItem(i, glassPane);
       }
     }
+    ItemStack air = new ItemStack(Material.AIR);
+    inventory.setItem(20, air);
+    inventory.setItem(29, air);
 
-    ItemStack heartChunk = new ItemBuilderManager(Material.SPIDER_EYE).setTitle("&c&lHeart Chunk").addLoreLine("&8Custom Item").toItemStack();
-    inventory.setItem(10, heartChunk);
-    inventory.setItem(11, heartChunk);
-    inventory.setItem(12, heartChunk);
-    inventory.setItem(19, heartChunk);
-    inventory.setItem(21, heartChunk);
-    inventory.setItem(28, heartChunk);
-    inventory.setItem(29, heartChunk);
-    inventory.setItem(30, heartChunk);
+    ItemStack leather = new ItemStack(Material.LEATHER);
+    inventory.setItem(10, leather);
+    inventory.setItem(11, leather);
+    inventory.setItem(12, leather);
 
-    ItemStack soulVial = new ItemBuilderManager(Material.GLASS_BOTTLE).setTitle("&b&lSoul Vial").addLoreLine("&8Custom Item").toItemStack();
-    inventory.setItem(20, soulVial);
+    ItemStack copper = new ItemStack(Material.COPPER_INGOT);
+    inventory.setItem(19, copper);
+    inventory.setItem(21, copper);
 
-    ItemStack heart = new ItemBuilderManager(Material.FERMENTED_SPIDER_EYE).setTitle("&c&lHeart").addLoreLine("&8Custom Item").toItemStack();
-    inventory.setItem(24, heart);
+    ItemStack tripwireHook = new ItemStack(Material.TRIPWIRE_HOOK);
+    inventory.setItem(28, tripwireHook);
+    inventory.setItem(30, tripwireHook);
+
+    ItemStack saddle = new ItemStack(Material.SADDLE);
+    inventory.setItem(24, saddle);
 
     ItemStack exit = new ItemBuilderManager(Material.RED_STAINED_GLASS_PANE).setTitle("&c&lBack").toItemStack();
     inventory.setItem(40, exit);
@@ -50,7 +57,7 @@ public class Heart implements Listener {
     if (event.getClickedInventory() == null) {
       return;
     }
-    if (event.getView().getTitle().equalsIgnoreCase(fix("&3&lRecipe: &c&lHeart"))) {
+    if (event.getView().getTitle().equalsIgnoreCase(fix("&3&lRecipe: &8Saddle"))) {
       event.setCancelled(true);
       Player player = (Player) event.getWhoClicked();
       if (event.getCurrentItem() == null) {
