+++ src/main/java/net/boybacks/superhardcore/recipes/items/craftable/vanilla/TotemOfUndying.java
@@ -1,6 +1,7 @@
-package net.boybacks.superhardcore.craftingrecipes.items;
+package net.boybacks.superhardcore.recipes.items.craftable.vanilla;
 
-import net.boybacks.superhardcore.craftingrecipes.handlers.onCraftingInventory;
+import lombok.Getter;
+import net.boybacks.superhardcore.recipes.handlers.onCraftingInventory;
 import net.boybacks.superhardcore.managers.ItemBuilderManager;
 import org.bukkit.Bukkit;
 import org.bukkit.Material;
@@ -13,34 +14,37 @@ import org.bukkit.inventory.ItemStack;
 
 import static net.boybacks.superhardcore.managers.ChatManager.fix;
 
-public class Crystal implements Listener {
+public class TotemOfUndying implements Listener {
+
+  @Getter static ItemStack totemOfUndying = new ItemStack(Material.TOTEM_OF_UNDYING);
 
   public static void onInventory(Player player) {
-    Inventory inventory = Bukkit.createInventory(player, 45, fix("&3&lRecipe: &d&lCrystal"));
+    Inventory inventory = Bukkit.createInventory(player, 45, fix("&3&lRecipe: &e&lTotem of Undying"));
     for (int i = 0; i < inventory.getSize(); i++) {
       if(inventory.getItem(i) == null || inventory.getItem(i).getType().equals(Material.AIR)) {
         ItemStack glassPane = new ItemBuilderManager(Material.WHITE_STAINED_GLASS_PANE).setTitle("&a").toItemStack();
         inventory.setItem(i, glassPane);
       }
     }
-
-    ItemStack redstone = new ItemStack(Material.REDSTONE);
-    inventory.setItem(10, redstone);
-    inventory.setItem(12, redstone);
-    inventory.setItem(19, redstone);
-    inventory.setItem(21, redstone);
-    inventory.setItem(28, redstone);
-    inventory.setItem(29, redstone);
-    inventory.setItem(30, redstone);
-
     ItemStack emerald = new ItemStack(Material.EMERALD);
-    inventory.setItem(11, emerald);
+    inventory.setItem(10, emerald);
+    inventory.setItem(12, emerald);
 
-    ItemStack amethyst = new ItemStack(Material.AMETHYST_SHARD);
-    inventory.setItem(20, amethyst);
+    ItemStack goldIngot = new ItemStack(Material.GOLD_INGOT);
+    inventory.setItem(28, goldIngot);
+    inventory.setItem(30, goldIngot);
 
-    ItemStack crystal = new ItemBuilderManager(Material.EMERALD).setTitle("&d&lCrystal").addLoreLine("&8Custom Item").toItemStack();
-    inventory.setItem(24, crystal);
+    ItemStack rippedHeart = new ItemBuilderManager(Material.FERMENTED_SPIDER_EYE).setTitle("&c&lRipped Heart").addLoreLine("&8Custom Item").addGlow().toItemStack();
+    inventory.setItem(20, rippedHeart);
+
+    ItemStack infusedCrystal = new ItemBuilderManager(Material.EMERALD).setTitle("&d&lInfused Crystal").addLoreLine("&8Custom Item").addGlow().toItemStack();
+    inventory.setItem(11, infusedCrystal);
+    inventory.setItem(19, infusedCrystal);
+    inventory.setItem(21, infusedCrystal);
+    inventory.setItem(29, infusedCrystal);
+
+    ItemStack totemOfUndying = new ItemBuilderManager(Material.TOTEM_OF_UNDYING).toItemStack();
+    inventory.setItem(24, totemOfUndying);
 
     ItemStack exit = new ItemBuilderManager(Material.RED_STAINED_GLASS_PANE).setTitle("&c&lBack").toItemStack();
     inventory.setItem(40, exit);
@@ -53,7 +57,7 @@ public class Crystal implements Listener {
     if (event.getClickedInventory() == null) {
       return;
     }
-    if (event.getView().getTitle().equalsIgnoreCase(fix("&3&lRecipe: &d&lCrystal"))) {
+    if (event.getView().getTitle().equalsIgnoreCase(fix("&3&lRecipe: &e&lTotem of Undying"))) {
       event.setCancelled(true);
       Player player = (Player) event.getWhoClicked();
       if (event.getCurrentItem() == null) {
@@ -64,7 +68,8 @@ public class Crystal implements Listener {
       }
     }
   }
-  /*
+
+    /*
 
  public static void onInventory(Player player) {
     Inventory inventory = Bukkit.createInventory(player, 45, fix("&3&lRecipe: "));
